+++ parrot/src/buffers/mod.rs
@@ -1,5 +1,7 @@
 pub mod vertex;
 pub mod index;
 pub mod uniform;
+pub mod depth;
+pub mod frame;
 
-pub use {vertex::VertexBuffer, index::IndexBuffer, uniform::UniformBuffer};
+pub use {vertex::VertexBuffer, index::IndexBuffer, uniform::UniformBuffer, depth::DepthBuffer, frame::FrameBuffer};