+++ parrot/src/device.rs
@@ -1,5 +1,5 @@
 use euclid::Size2D;
-use wgpu::util::DeviceExt;
+use wgpu::{util::DeviceExt, TextureFormat, TextureUsages};
 use crate::{
     vertex::VertexLayout,
     transform::ScreenSpace,
@@ -10,7 +10,7 @@ use crate::{
     buffers::{
         vertex::VertexBuffer,
         index::IndexBuffer,
-        uniform::UniformBuffer,
+        uniform::UniformBuffer, DepthBuffer, FrameBuffer
     },
     texture::Texture,
     sampler::Sampler,
@@ -27,6 +27,7 @@ pub struct Device {
     pub queue: wgpu::Queue,
     /// Wrapper around the surface (if there is one)
     pub surface: Option<wgpu::Surface>,
+    /// Size of the surface
     size: euclid::Size2D<u32, ScreenSpace>,
 }
 
@@ -91,6 +92,7 @@ impl Device {
 
     /// Create a shader given a [`crate::shader::ShaderFile`]
     pub fn create_shader(&self, source: ShaderFile, name: Option<&str>) -> Shader {
+        log::info!("Creating shader >> Name: {:?}", name);
         match source {
             ShaderFile::Spirv(bytes) => self.create_sprv_shader(bytes, name),
             ShaderFile::Wgsl(s) => self.create_wgsl_shader(s, name),
@@ -119,19 +121,23 @@ impl Device {
         }
     }
 
-    pub fn create_vertex_buffer<T: bytemuck::Pod>(&self, verticies: &[T], name: Option<&str>) -> VertexBuffer {
+    pub fn create_vertex_buffer<T: bytemuck::Pod>(&self, vertices: &[T], name: Option<&str>) -> VertexBuffer {
+        log::info!("Created vertex buffer >> Name: {:?}", name);
         VertexBuffer {
-            wgpu: self.create_buffer_from_slice(verticies, wgpu::BufferUsages::VERTEX, 
+            wgpu: self.create_buffer_from_slice(vertices, wgpu::BufferUsages::VERTEX | wgpu::BufferUsages::COPY_DST, 
                 name),
-            size: (verticies.len() * std::mem::size_of::<T>()) as u32
+            size: (vertices.len() * std::mem::size_of::<T>()) as u32,
+            name: name.map(|s| s.to_string()),
         }
     }
 
     pub fn create_index_buffer(&self, indicies: &[u16], name: Option<&str>) -> IndexBuffer {
-        let index_buf = self.create_buffer_from_slice(indicies, wgpu::BufferUsages::INDEX, name);
+        log::info!("Created index buffer >> Name: {:?}", name);
+        let index_buf = self.create_buffer_from_slice(indicies, wgpu::BufferUsages::INDEX | wgpu::BufferUsages::COPY_DST, name);
         IndexBuffer {
             wgpu: index_buf,
-            elements: indicies.len() as u32,
+            size: indicies.len() as u32,
+            name: name.map(|s| s.to_string())
         }
     }
 
@@ -139,6 +145,7 @@ impl Device {
     where
         T: bytemuck::Pod + 'static + Copy
     {
+        log::info!("Created uniform buffer");
         UniformBuffer {
             size: std::mem::size_of::<T>(),
             count: buf.len(),
@@ -148,6 +155,36 @@ impl Device {
                 usage: wgpu::BufferUsages::UNIFORM | wgpu::BufferUsages::COPY_DST,
             }),
         }
+    }
+
+    /// Create a depth buffer
+    pub fn create_depth_buffer(&self, size: Size2D<u32, ScreenSpace>, sample_count: u32, name: Option<&str>) -> DepthBuffer {
+        log::info!("Created depth buffer");
+        let format = DepthBuffer::FORMAT;
+        let extent = wgpu::Extent3d {
+            width: size.width,
+            height: size.height,
+            depth_or_array_layers: 1,
+        };
+
+        let wgpu = self.wgpu.create_texture(&wgpu::TextureDescriptor {
+            size: extent,
+            label: name,
+            mip_level_count: 1,
+            sample_count,
+            dimension: wgpu::TextureDimension::D2,
+            format,
+            usage: wgpu::TextureUsages::COPY_DST | wgpu::TextureUsages::RENDER_ATTACHMENT,
+        });
+        let view = wgpu.create_view(&wgpu::TextureViewDescriptor::default());
+
+        DepthBuffer { texture: Texture {
+            wgpu,
+            view,
+            extent,
+            format,
+            size,
+        }}
     }
 
     pub fn create_buffer_from_slice<T: bytemuck::Pod> (
@@ -170,18 +207,20 @@ impl Device {
         format: wgpu::TextureFormat,
         usage: wgpu::TextureUsages,
         name: Option<&str>,
+        sample_count: u32,
     ) -> Texture {
+        log::info!("Creating texture >> Name: {:?}", name);
         let texture_extent = wgpu::Extent3d {
             width: size.width,
             height: size.height,
-            depth_or_array_layers: 0,
+            depth_or_array_layers: 1,
         };
 
         let texture = self.wgpu.create_texture( &wgpu::TextureDescriptor {
             label: name,
             size: texture_extent,
             mip_level_count: 1,
-            sample_count: 1,
+            sample_count,
             dimension: wgpu::TextureDimension::D2,
             format,
             usage,
@@ -199,6 +238,7 @@ impl Device {
     }
 
     pub fn create_sampler(&self, mag_filter: wgpu::FilterMode, min_filter: wgpu::FilterMode, name: Option<&str>) -> Sampler {
+        log::info!("Creating sampler >> Name: {:?}", name);
         Sampler {
             wgpu: self.wgpu.create_sampler( &wgpu::SamplerDescriptor{
                 label: name,
@@ -217,7 +257,51 @@ impl Device {
         }
     }
 
+    pub fn create_frame_buffer(&self, size: Size2D<u32, ScreenSpace>, format: TextureFormat, sample_count: u32, name: Option<&str>, depth: bool) -> FrameBuffer {
+        log::info!("Creating frame buffer >> Name: {:?} || Depth: {}", name, depth);
+        let extent = wgpu::Extent3d {
+            width: size.width,
+            height: size.height,
+            depth_or_array_layers: 1
+        };
+        let texture = self.wgpu.create_texture(&wgpu::TextureDescriptor {
+            size: extent,
+            mip_level_count: 1,
+            sample_count,
+            dimension: wgpu::TextureDimension::D2,
+            format,
+            usage: TextureUsages::TEXTURE_BINDING | TextureUsages::COPY_DST | TextureUsages::COPY_SRC | TextureUsages::RENDER_ATTACHMENT,
+            label: name
+        });
+        let view = texture.create_view(&wgpu::TextureViewDescriptor::default());
+
+        if depth {
+            FrameBuffer {
+                texture: Texture {
+                    wgpu: texture,
+                    view,
+                    extent,
+                    format,
+                    size
+                },
+                depth: Some(self.create_depth_buffer(size, sample_count, name))
+            }
+        } else {
+            FrameBuffer {
+                texture: Texture {
+                    wgpu: texture,
+                    view,
+                    extent,
+                    format,
+                    size
+                },
+                depth: None
+            }
+        }        
+    }
+
     pub fn create_binding_group_layout(&self, index: u32, slots: &[Binding], name: Option<&str>) -> BindingGroupLayout {
+        log::info!("Creating bind group layout >> Name: {:?} || Index: {:?}", name, index);
         let mut bindings = Vec::new();
 
         for s in slots {
@@ -237,6 +321,7 @@ impl Device {
     }
 
     pub fn create_binding_group(&self, layout: &BindingGroupLayout, binds: &[&dyn Bind], name: Option<&str>) -> BindingGroup {
+        log::info!("Creating binding >> Name: {:?}", name);
         assert_eq!(binds.len(), layout.size, "Layout slot doesn't match bindings");
 
         let mut bindings = Vec::new();
@@ -280,14 +365,20 @@ impl Device {
     /// Updates a vertex buffer
     pub fn update_vertex_buffer<T: bytemuck::Pod + Copy + 'static>(&self, vertices: &[T], buf: &mut VertexBuffer) {
         self.queue.write_buffer(&buf.wgpu, 0, bytemuck::cast_slice(vertices));
-        // Update the size of it.
-        buf.size = (vertices.len() * std::mem::size_of::<T>()) as u32
     }
 
     /// Update a index buffer
-    pub fn update_index_buffer(&self, indicies: &[u16], buf: &mut IndexBuffer) {
-        self.queue.write_buffer(&buf.wgpu, 0, bytemuck::cast_slice(indicies));
-        buf.elements = indicies.len() as u32
+    pub fn update_index_buffer(&self, mut indicies: Vec<u16>, buf: &mut IndexBuffer) {
+        // Get the alignment
+        let alignment = wgpu::COPY_BUFFER_ALIGNMENT as usize / std::mem::size_of::<u16>();
+        let fraction = indicies.len() % alignment;
+        // Extend the index buffer so its aligned
+        if fraction > 0 {
+            indicies.extend(std::iter::repeat(0).take(alignment - fraction));
+        }
+
+        // Update the buffer
+        self.queue.write_buffer(&buf.wgpu, 0, bytemuck::cast_slice(indicies.as_slice()));
     }
 
     /// Create a pipeline
@@ -296,10 +387,10 @@ impl Device {
         pipeline_layout: PipelineLayout,
         vertex_layout: VertexLayout,
         blending: Blending,
-        shader: &Shader,
+        shader: Shader,
         tex_format: wgpu::TextureFormat,
         multisample: wgpu::MultisampleState,
-        name: String
+        name: Option<&str>
     ) -> Pipeline {
         let vertex_attrs = vertex_layout.to_wgpu();
         let mut b_layouts = Vec::new();
@@ -309,7 +400,7 @@ impl Device {
         }
 
         let layout = &self.wgpu.create_pipeline_layout(&wgpu::PipelineLayoutDescriptor {
-            label: Some(&format!("{} pipeline layout", &name)),
+            label: name,
             bind_group_layouts: b_layouts.as_slice(),
             push_constant_ranges: &[],
         });
@@ -334,10 +425,103 @@ impl Device {
             write_mask: wgpu::ColorWrites::ALL,
         }];
 
-        let n = format!("{} pipeline", &name);
+        let desc = wgpu::RenderPipelineDescriptor {
+            label: name,
+            layout: Some(layout),
+            vertex: wgpu::VertexState {
+                module: &shader.wgpu,
+                entry_point: "vs_main",
+                buffers: &[vertex_attrs],
+            },
+            primitive: wgpu::PrimitiveState {
+                topology: wgpu::PrimitiveTopology::TriangleList,
+                strip_index_format: None,
+                front_face: wgpu::FrontFace::Ccw,
+                cull_mode: None,
+                polygon_mode: wgpu::PolygonMode::Fill,
+                unclipped_depth: false,
+                conservative: false,
+            },
+            depth_stencil: Some(wgpu::DepthStencilState {
+                format: DepthBuffer::FORMAT,
+                depth_write_enabled: true,
+                depth_compare: wgpu::CompareFunction::LessEqual,
+                stencil: wgpu::StencilState {
+                    front: wgpu::StencilFaceState::IGNORE,
+                    back: wgpu::StencilFaceState::IGNORE,
+                    read_mask: 0,
+                    write_mask: 0,
+                },
+                bias: wgpu::DepthBiasState {
+                    constant: 0,
+                    slope_scale: 0.,
+                    clamp: 0.,
+                }
+            }),
+            multisample,
+            fragment: Some(wgpu::FragmentState {
+                module: &shader.wgpu,
+                entry_point: "fs_main",
+                targets: &targets,
+            }),
+            multiview: None,
+        };
+
+        let wgpu = self.wgpu.create_render_pipeline(&desc);
+
+        Pipeline {
+            layout: pipeline_layout,
+            vertex_layout,
+            wgpu,
+        }
+    }
+
+    /// Create a pipeline without a depth buffer
+    pub fn create_pipeline_no_depth(
+        &self,
+        pipeline_layout: PipelineLayout,
+        vertex_layout: VertexLayout,
+        blending: Blending,
+        shader: Shader,
+        tex_format: wgpu::TextureFormat,
+        multisample: wgpu::MultisampleState,
+        name: Option<&str>
+    ) -> Pipeline {
+        let vertex_attrs = vertex_layout.to_wgpu();
+        let mut b_layouts = Vec::new();
+
+        for s in pipeline_layout.b_layouts.iter() {
+            b_layouts.push(&s.wgpu);
+        }
+
+        let layout = &self.wgpu.create_pipeline_layout(&wgpu::PipelineLayoutDescriptor {
+            label: name,
+            bind_group_layouts: b_layouts.as_slice(),
+            push_constant_ranges: &[],
+        });
+
+        let (src_factor, dst_factor, operation) = blending.as_wgpu();
+
+        // I like your funny words magic man
+        let targets = [wgpu::ColorTargetState {
+            format: tex_format,
+            blend: Some(wgpu::BlendState {
+                color: wgpu::BlendComponent {
+                    src_factor,
+                    dst_factor,
+                    operation
+                },
+                alpha: wgpu::BlendComponent {
+                    src_factor,
+                    dst_factor,
+                    operation
+                },
+            }),
+            write_mask: wgpu::ColorWrites::ALL,
+        }];
 
         let desc = wgpu::RenderPipelineDescriptor {
-            label: Some(&n),
+            label: name,
             layout: Some(layout),
             vertex: wgpu::VertexState {
                 module: &shader.wgpu,
@@ -370,5 +554,20 @@ impl Device {
             vertex_layout,
             wgpu,
         }
+    }
+
+    pub fn create_render_bundle_encoder(&self, format: wgpu::TextureFormat, name: Option<&str>, sample_count: u32) -> wgpu::RenderBundleEncoder {
+        log::info!("Creating render bundle encoder >> Name: {:?}", name);
+        self.wgpu.create_render_bundle_encoder(&wgpu::RenderBundleEncoderDescriptor {
+            label: name,
+            depth_stencil: Some(wgpu::RenderBundleDepthStencil {
+                format: DepthBuffer::FORMAT,
+                depth_read_only: false,
+                stencil_read_only: false,
+            }),
+            sample_count,
+            color_formats: &[format],
+            multiview: None,
+        })
     }
 }