+++ parrot/CHANGELOG.md
@@ -43,3 +43,32 @@ This keeps note of the API changes so I don't have to stare at commit messages a
 - The prepare function for plumber now works for multiple uniform buffers
 
 - Update buffer functions now properly update a buffers size
+
+## Version 0.2.0
+
+Finally, I created some examples and tests and discovered: Everything doesn't work. D: So I went on a quest to fix it all.
+
+This update also brings multisampling and depth textures into parrot and (hopefully) makes the update buffer functions actually work. As I have just implemented depth buffers and multisampling be aware there's probably going to be bugs.
+Also: **EXAMPLES WOOOOOOOOOOOOOOOOOOOOO**, now you can see parrot in action
+
+### Changes
+
+- Updating buffer functions now either update or create a new buffer big enough to fit the data
+
+- Removed the requirement to specify the vertex type your using
+
+- Fixed a lot of bugs
+
+## New
+
+- Added depth textures
+
+- Added multisampling
+
+- Added examples and improved documentation
+
+- Added custom pipelines
+
+- Added frame buffers
+
+- Added basic logging